data_layer/requirements.txt
python-dotenv
psycopg2-binary
sqlalchemy >= 1.1.2
Flask-Migrate
